+++ src/config/keys.rs
@@ -112,6 +112,7 @@ impl Default for KeyBindings {
         insert_binding!(general, "K", Command::UpDirectory);
         insert_binding!(general, "shift-up", Command::UpDirectory);
         insert_binding!(general, "f", Command::LoadSaveFile);
+        insert_binding!(general, "shift-alt-f", Command::LoadRandomSaveFile);
         insert_binding!(general, "ctrl-f", Command::LoadActiveSaveFile);
         insert_binding!(general, "F", Command::MarkSaveFile);
         insert_binding!(general, "i", Command::ImportSaveFile);